fix: price field crashing when value exceeds MAX_LONG

+++ composeApp/src/commonMain/kotlin/br/com/noartcode/theprice/data/local/dao/BillDao.kt
@@ -35,7 +35,7 @@ interface BillDao {
         id:Long,
         name:String,
         description:String?,
-        price:Int,
+        price:Long,
         type:String,
         status:String,
         billingStartDate:Long
