+++ composeApp/src/commonMain/kotlin/br/com/noartcode/theprice/domain/model/Bill.kt
@@ -4,7 +4,7 @@ data class Bill(
     val id:Long = 0,
     val name:String = "",
     val description:String? = null,
-    val price:Int = 0,
+    val price:Long = 0,
     val type: Type = Type.MONTHLY,
     val status: Status = Status.ACTIVE,
     val billingStartDate: DayMonthAndYear,
